# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": true,
          "error": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "error"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
...
      "PromptCreateRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Request to create a new prompt.",
        "properties": {
          "content": {
            "type": "string"
          },
          "description": {
            "default": "",
            "type": "string"
          },
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "content",
          "name"
        ],
        "title": "PromptCreateRequest",
        "type": "object"
      },
      "PromptMutationResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Response for prompt create / update / delete mutations.",
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "ok": {
            "type": "boolean"
          },
          "prompt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuickPrompt"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ok"
        ],
        "title": "PromptMutationResponse",
        "type": "object"
      },
      "PromptUpdateRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Request to update an existing prompt.",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "title": "PromptUpdateRequest",
        "type": "object"
      },
      "PromptsListResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Response for listing all prompts.",
        "properties": {
          "prompts": {
            "items": {
              "$ref": "#/components/schemas/QuickPrompt"
            },
            "type": "array"
          }
        },
        "required": [
          "prompts"
        ],
        "title": "PromptsListResponse",
        "type": "object"
      },
...
      "QuickPrompt": {
        "description": "Quick prompt entry.",
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "content",
          "created_at",
          "description",
          "id",
          "name",
          "updated_at"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "allthecodes Web API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/prompts": {
      "get": {
        "operationId": "PromptsList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptsListResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      },
      "post": {
        "operationId": "PromptsCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
    "/api/prompts/{id}": {
      "delete": {
        "operationId": "PromptsDelete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      },
      "get": {
        "operationId": "PromptsDetail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      },
      "patch": {
        "operationId": "PromptsUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptMutationResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
...
  }
}